requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
MouseInfo==0.1.3
mss==9.0.1
outcome==1.2.0
Pillow==10.0.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytweening==1.0.7
reportlab==4.0.4
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
wsproto==1.2.0
